Visual + Design Rework
Made my name look nicer and bigger. Made a new background to spice
things up. Added a references popup to the References link, which allows
for multiple sheets/letters to be linked. Moved the footer block to hug
the bottom of the screen. Made the whole site resize nicely and work on
really big screens and mobile. Moved the project labels into the tiles
themselves. Changed the body font.